requirements.txt
streamlit
python-docx
PyMuPDF
nltk
transformers
torch
